requirements.txt
opencv_python==4.5.2.52
numpy==1.20.1
streamlit==0.82.0
fastai==2.3.1
Pillow==8.2.0
